requirements.txt
Flask==2.3.3
python-nmap==0.7.1
ping3==4.0.4
pandas==2.0.3
openpyxl==3.1.2
python-docx==0.8.11
gunicorn==21.2.0
Flask==2.3.3
python-nmap==0.7.1
python-docx==0.8.11